IPc 将ipc代码移进IpcManager GetOrderStatus 移除UI操作 TravelManager GetTravelingOrder 是获取返回订单的，所以暴露给 IpcManager
